+++ eventsubscribers/health_checkers_test.go
@@ -56,7 +56,7 @@ func TestMojangBatchUuidsProviderChecker(t *testing.T) {
 		checker := MojangBatchUuidsProviderResponseChecker(d, time.Millisecond)
 		assert.Nil(t, checker(context.Background()))
 	})
-	//
+
 	t.Run("when no error occurred", func(t *testing.T) {
 		d := dispatcher.New()
 		checker := MojangBatchUuidsProviderResponseChecker(d, time.Millisecond)
@@ -105,5 +105,42 @@ func TestMojangBatchUuidsProviderQueueLengthChecker(t *testing.T) {
 		if assert.Error(t, checkResult) {
 			assert.Equal(t, "the maximum number of tasks in the queue has been exceeded", checkResult.Error())
 		}
+	})
+}
+
+func TestMojangApiTexturesProviderResponseChecker(t *testing.T) {
+	t.Run("empty state", func(t *testing.T) {
+		d := dispatcher.New()
+		checker := MojangApiTexturesProviderResponseChecker(d, time.Millisecond)
+		assert.Nil(t, checker(context.Background()))
+	})
+
+	t.Run("when no error occurred", func(t *testing.T) {
+		d := dispatcher.New()
+		checker := MojangApiTexturesProviderResponseChecker(d, time.Millisecond)
+		d.Emit("mojang_textures:mojang_api_textures_provider:after_request",
+			"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+			&mojang.SignedTexturesResponse{},
+			nil,
+		)
+		assert.Nil(t, checker(context.Background()))
+	})
+
+	t.Run("when error occurred", func(t *testing.T) {
+		d := dispatcher.New()
+		checker := MojangApiTexturesProviderResponseChecker(d, time.Millisecond)
+		err := errors.New("some error occurred")
+		d.Emit("mojang_textures:mojang_api_textures_provider:after_request", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", nil, err)
+		assert.Equal(t, err, checker(context.Background()))
+	})
+
+	t.Run("should reset value after passed duration", func(t *testing.T) {
+		d := dispatcher.New()
+		checker := MojangApiTexturesProviderResponseChecker(d, 20*time.Millisecond)
+		err := errors.New("some error occurred")
+		d.Emit("mojang_textures:mojang_api_textures_provider:after_request", "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa", nil, err)
+		assert.Equal(t, err, checker(context.Background()))
+		time.Sleep(40 * time.Millisecond)
+		assert.Nil(t, checker(context.Background()))
 	})
 }