+++ mojangtextures/mojang_api_textures_provider.go
@@ -13,7 +13,7 @@ type MojangApiTexturesProvider struct {
 func (ctx *MojangApiTexturesProvider) GetTextures(uuid string) (*mojang.SignedTexturesResponse, error) {
 	ctx.Emit("mojang_textures:mojang_api_textures_provider:before_request", uuid)
 	result, err := uuidToTextures(uuid, true)
-	ctx.Emit("mojang_textures:mojang_api_textures_provider:after_request", result, err)
+	ctx.Emit("mojang_textures:mojang_api_textures_provider:after_request", uuid, result, err)
 
 	return result, err
 }